backend/pyproject.toml
[build-system]
requires = ["setuptools>=67", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "kanban-lite-backend"
version = "0.1.0"
description = "Kanban Lite API"
requires-python = ">=3.12"
dependencies = [
  "fastapi==0.110.2",
  "uvicorn[standard]==0.29.0",
  "sqlalchemy==2.0.29",
  "psycopg[binary]==3.1.18",
  "alembic==1.13.1",
  "pydantic-settings==2.2.1",
  "python-dotenv==1.0.1"
]

[project.optional-dependencies]
dev = [
  "pytest==8.2.0",
  "httpx==0.27.0",
  "ruff==0.4.2"
]

[tool.pytest.ini_options]
pythonpath = ["app"]
testpaths = ["tests"]
filterwarnings = ["ignore::DeprecationWarning"]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.ruff.lint.isort]
known-first-party = ["app"]
profile = "black"

[tool.setuptools]
packages = ["app"]
